+++ src/Polly.Testing/README.md
@@ -2,6 +2,10 @@
 
 This package exposes APIs and utilities that can be used to assert on the composition of resilience pipelines.
 
+See <https://www.pollydocs.org/advanced/testing> for more details.
+
+## Usage
+
 <!-- snippet: get-pipeline-descriptor -->
 ```cs
 // Build your resilience pipeline.
@@ -11,19 +15,20 @@ ResiliencePipeline pipeline = new ResiliencePipelineBuilder()
         MaxRetryAttempts = 4
     })
     .AddTimeout(TimeSpan.FromSeconds(1))
-    .ConfigureTelemetry(NullLoggerFactory.Instance)
     .Build();
 
-// Retrieve inner strategies.
+// Retrieve the descriptor.
 ResiliencePipelineDescriptor descriptor = pipeline.GetPipelineDescriptor();
 
-// Assert the composition.
+// Check the pipeline's composition with the descriptor.
 Assert.Equal(2, descriptor.Strategies.Count);
 
+// Verify the retry settings.
 var retryOptions = Assert.IsType<RetryStrategyOptions>(descriptor.Strategies[0].Options);
 Assert.Equal(4, retryOptions.MaxRetryAttempts);
 
-var timeoutOptions = Assert.IsType<TimeoutStrategyOptions>(descriptor.Strategies[0].Options);
+// Confirm the timeout settings.
+var timeoutOptions = Assert.IsType<TimeoutStrategyOptions>(descriptor.Strategies[1].Options);
 Assert.Equal(TimeSpan.FromSeconds(1), timeoutOptions.Timeout);
 ```
 <!-- endSnippet -->
